refactor(runtime-host): keep owner policy outside protocol
Generated-by: Codex
